requirements.txt
# Generated 2025-02-17 10:17:49.377226
toga-web~=0.4.7
